Added pretty sales for node lists.
TODO: update editor via ajax, so that when the user selects a node the page is not reloaded (this causes the .selected to be lost), thus retaining the correct style for the selected node.

Another solution would be to apply CSS depending on the currently displayed node (currentSource and currentRelated).